requirements.txt
carvekit==4.1.2
diffusers==0.29.2
fire==0.7.0
lpips==0.1.4
matplotlib==3.9.4
megfile==3.1.3
numpy==2.2.0
opencv_contrib_python==4.9.0.80
opencv_contrib_python_headless==4.9.0.80
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.82
Pillow==11.0.0
Requests==2.32.3
seaborn==0.13.2
segment_anything==1.0
skimage==0.0
submitit==1.5.2
timm==1.0.12
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.5
transformers==4.44.0
